Move 'SKIP' to launch_utils and work through all remaining TODOs
