+++ ros2_ws/src/rcdt_launch/rcdt_launch/platforms/vehicle.py
@@ -160,7 +160,7 @@ class Vehicle(Platform):
             launch_descriptions.append(self.create_nav2_launch())
         return launch_descriptions
 
-    def create_map_link(self) -> Node | None:  # TODO: naar classen zelf opsplitsen!
+    def create_map_link(self) -> Node | None:
         """Create a static_transform_publisher node that links the platform to the map.
 
         If the platform is a Vehicle that is using SLAM or navigation, None is returned.
